% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_irsz_tsz_crosswalk}
\alias{get_irsz_tsz_crosswalk}
\title{Get a postal code to settlement ID crosswalk}
\usage{
get_irsz_tsz_crosswalk(year = c("2018"))
}
\arguments{
\item{year}{Version of the crosswalk table to be used. Defaults to
2018.}
}
\value{
A tibble
}
\description{
Aggregate a settlement ID to postal code crosswalk table by the unique
primary settlement (\code{torzsszam_fo_telepules}) to get an unambiguous
postal code to settlement ID crosswalk.
}
\details{
Note that currently this package includes just a single crosswalk
  table, \code{\link{irsz_2018}}.
}
