% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{irsz_tsz_crosswalk_aggreage}
\alias{irsz_tsz_crosswalk_aggreage}
\title{Helper function for \code{\link{get_irsz_tsz_crosswalk}}}
\usage{
irsz_tsz_crosswalk_aggreage(df)
}
\arguments{
\item{df}{A settlement ID to postal code crosswalk table to aggregate.}
}
\description{
Helper function for \code{\link{get_irsz_tsz_crosswalk}}
}
\keyword{internal}
