% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teroszt.R
\docType{data}
\name{kozighatarok_2018}
\alias{kozighatarok_2018}
\title{Geospatial data of Hungarian administrative divisions from OpenStreetMap}
\format{A Simple Features data frame.

\describe{
  \item{NAME}{Name of the administrative division.}
  \item{ADMIN_LEVE}{OpenStreetMap \code{admin_level} values, see \url{https://wiki.openstreetmap.org/wiki/Tag:boundary\%3Dadministrative#10_admin_level_values_for_specific_countries}.}
  \item{ADMIN_NAME}{Label for \code{ADMIN_LEVE}.}
  \item{CODE}{IDs used by HCSO for administrative divisions.}
  \item{geometry}{Geospatial information.}
}}
\source{
OpenStreetMap \url{https://data2.openstreetmap.hu/hatarok/}, licenced
  under CC BY-SA \url{https://www.openstreetmap.org/copyright} ©
  OpenStreetMap contributors. Downloaded on 13th September 2019.
}
\usage{
kozighatarok_2018
}
\description{
A data frame with geographic boundaries for all Hungarian administrative
divisions based on OpenStreetMap data. Adopted for compatibility with HCSO
territorial code system.
}
\keyword{datasets}
