% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teroszt.R
\docType{data}
\name{nav_igazgatosagi_kodok}
\alias{nav_igazgatosagi_kodok}
\title{NAV igazgatósági kódok}
\format{Egy data frame 199 sorral, amely minden járáshoz megadja az illetékes NAV
igazgatóság kódját.  A 999-es, fiktív járáshoz lett sorolva a két kiemelt
igazgatóság.

\describe{
  \item{nav_ig_nev}{Az igazgatóság neve}
  \item{nav_ig_tarsas}{Igazgatóság kódja, társas adóalany esetén}
  \item{nav_ig_egyeni}{Igazgatóság kódja, egyéni adóalany esetén}
  \item{megye}{Megye kódja (ld. TSZJ)}
  \item{jaras}{Járás kódja (ld. TSZJ)}
}}
\source{
(\url{http://www.nav.gov.hu/nav/adatbazisok/adatbleker/afaalanyok/afaalany_taj.html}).
A hozzáférés dátuma 2019. április 10.
}
\usage{
nav_igazgatosagi_kodok
}
\description{
A Nemzeti Adó- és Vámhivatal által használt területi igazgatósági kódok és a
hozzájuk kapcsolódó megye- és járáskódok a területi számrendszerből.  Ezek az
igazgatósági kódok az adószám utolsó két számjegyei is.
}
\keyword{datasets}
