% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teroszt.R
\docType{package}
\name{teroszt-package}
\alias{teroszt}
\alias{teroszt-package}
\title{teroszt: Magyarországon használt területi osztályozások adattáblái}
\description{
Ez az R csomag a KSH Területi számjelrendszerének
    <http://www.ksh.hu/teruleti_szamjel_menu>,
    Helységnévtárának <http://www.ksh.hu/apps/hntr.main>, illetve a postai
    irányítószámok <https://www.posta.hu/szolgaltatasok/iranyitoszam-kereso>
    KSH-s településazonosítókhoz és a NAV által használt igazgatósági kódok
    <http://www.nav.gov.hu/nav/adatbazisok/adatbleker/afaalanyok/afaalany_taj.html>
    területi osztályozásokhoz kapcsolt tábláit tartalmazza.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/svraka/teroszt}
  \item Report bugs at \url{https://github.com/svraka/teroszt/issues}
}

}
\author{
\strong{Maintainer}: András Svraka \email{svraka.andras@gmail.com}

}
\keyword{internal}
