% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teroszt.R
\docType{data}
\name{tsz_2018}
\alias{tsz_2018}
\title{2018 edition of the territorial code system}
\format{
A tibble with 3,200 rows, for each settlement that has an ID
  number (\sQuote{településazonosító törzsszám}) with codes for the
  classifications in the territorial code system, along with labels
  for the main classifications. For a detailed explanation of the
  variables see HCSO's documentation.
}
\source{
The present content was prepared using the territorial
  classification system of the Hungarian Central Statistical Office
  (\url{https://www.ksh.hu/tszJ_eng_menu}). What is included in
  this package is the intellectual product solely of this package's
  authors. Downloaded at 4 April 2019.
}
\usage{
tsz_2018
}
\description{
A dataset with the 2018 edition of the Hungarian Central
Statistical Office's territorial code system (\sQuote{területi
számjelrendszer}).
}
\keyword{datasets}
