% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teroszt.R
\docType{data}
\name{hnt_telepulesreszek_2018}
\alias{hnt_telepulesreszek_2018}
\title{Településrészek a Helységnévtár 2018-as kiadásából}
\format{Egy data frame 13 558 sorral, amely a Helységnévtárban szereplő
összes település településrészével az alábbi változókkal:

\describe{
  \item{torzsszam}{A település KSH-s településazonosító törzsszáma}
  \item{telepules}{Település neve}
  \item{telepulesresz}{Településrész neve}
  \item{telepulesresz_jelleg}{Településrész jellege (központi belterület,
    belterület, külterület)}
  \item{irsz}{Irányítószám (nem teljes körű, hiányoznak az utcajegyzék
    alapján meghatározott irányítószámok)}
  \item{kulterulet_jellege}{Külterület jellege a KSH módszertana szerint}
  \item{telepulesresz_tavolsaga_kozponti_belterulettol}{A településrész
    távolsága a központi belterülettől (km)}
  \item{nepszamlalasi_lakonepesseg}{A népszámlálási lakónépesség (fő)}
  \item{lakasok_szama}{A lakások száma (db)}
  \item{lakott_egyeb_lakoegysegek_szama}{Lakott egyéb lakóegységek száma
    (db)}
}}
\source{
A jelen tartalom a Központi Statisztikai Hivatal Helységnévtára
(\url{http://www.ksh.hu/apps/hntr.main}) 2018-as kiadásának felhasználásával
készült. A csomagban foglaltak kizárólag Svraka András, mint szerző szellemi
termékei.  A letöltés dátuma 2019. április 5.
}
\usage{
hnt_telepulesreszek_2018
}
\description{
Ez a tábla a KSH Helységnévtárának 2018-as kiadásából tartalmazza a
településrészeket.
}
\keyword{datasets}
