% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teroszt.R
\docType{data}
\name{irsz_2018}
\alias{irsz_2018}
\title{Irányítószámok és települések (2018)}
\format{Egy data frame 3859 sorral, amely minden, a Helységnévtárban és a Posta
nyilvántartásában szereplő irányítószámhoz megadja a települést.

\describe{
  \item{torzsszam}{A település KSH-s településazonosító törzsszáma.}
  \item{telepules}{Település neve.}
  \item{irsz}{Irányítószám.}
  \item{csak_kulterulet}{Vannak olyan irányítószámok, amelyek több
    településhez is tartoznak.  Ez az oszlop azt jelöli, hogy az adott
    irányítószám és település páros csak külterületként fordul-e elő.}
}}
\source{
A jelen tartalom a Központi Statisztikai Hivatal Helységnévtára
(\url{http://www.ksh.hu/apps/hntr.main}) 2018-as kiadásának felhasználásával
készült.  A csomagban foglaltak kizárólag Svraka András, mint szerző szellemi
termékei.  A letöltés dátuma 2019. április 4.

Az irányítószámok forrása a Magyar Posta Zrt. honlapján közzétett
\dQuote{Magyarországi postai irányítószámok}
(\url{https://www.posta.hu/szolgaltatasok/iranyitoszam-kereso}).  A letöltés
dátuma 2019. április 5.
}
\usage{
irsz_2018
}
\description{
A Magyar Posta nyilvántartása szerinti összes irányítószám, a kapcsolódó
település neve a KSH helységnévtárából hozzákapcsolva a településazonosító
törzsszámokat.
}
\keyword{datasets}
