% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teroszt.R
\docType{data}
\name{irsz_2018}
\alias{irsz_2018}
\title{Postal codes and settlements (2018)}
\format{
A tibble with 3,859 rows, which gives a crosswalk between
  all the valid postal codes in the Hungarian Postal Service's
  database and settlement IDs.

\describe{
  \item{torzsszam}{The settlement's HCSO ID number
    (\sQuote{településazonosító törzsszám})}
  \item{telepules}{Name of the settlement}
  \item{irsz}{Postal code}
  \item{csak_kulterulet}{Some settlements share postal codes. This
    column flags if a postal code within a settlement only covers area
    outside a built-up area (\sQuote{külterület}). This is intended
    to help classifying postal codes into administrative divisions in
    case the postal code is ambiguous.}
}
}
\source{
The present content was prepared using the Detailed
  Gazetteer of the Hungarian Central Statistical Office
  (\url{http://www.ksh.hu/apps/hntr.main?p_lang=EN}). What is
  included in this package is the intellectual product solely of
  this package's authors. Downloaded at 5 April 2019.

The source of the postal codes is Hungarian Postal Service's
(\sQuote{Magyar Posta Zrt.}) website:
(\url{https://www.posta.hu/szolgaltatasok/iranyitoszam-kereso}).
Downloaded at 5 April 2019.
}
\usage{
irsz_2018
}
\description{
A crosswalk table connecting Hungarian Postal Service's postal
codes (\sQuote{\emph{ir}ányító\emph{sz}ám}) and the HCSO's
settlement IDs.
}
\details{
Note that postal codes cross district and county level,
  and even regional boundries. Classifying postal codes into
  administrative divsions requires further data cleaning at this
  time.
}
\keyword{datasets}
