% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teroszt.R
\docType{data}
\name{irsz_2018}
\alias{irsz_2018}
\title{Postal codes and settlements (2018)}
\format{
A tibble with 4,405 rows, which gives a crosswalk between
  all the valid postal codes in the Hungarian Postal Service's
  database, and settlement IDs for all settlements that share a
  postal code.

\describe{
  \item{torzsszam}{The settlement's HCSO ID number
    (\sQuote{településazonosító törzsszám})}
  \item{telepules}{Name of the settlement}
  \item{irsz}{Postal code}
  \item{torzsszam_fo_telepules}{Settlement ID number for the primary
    settlement using a postal code. As settlements
    can share postal codes, a raw crosswalk between \code{irsz_2018}
    and \link{tsz_2018} is ambiguous. Most overlaps happen in
    parts of settlements with low-populations in non-central, or
    even non--build-up areas. This column contains a unique settlement ID
    for each postal code belonging to the settlement most likely to be
    representative of that postal code. Classification is
    based on borough types (preferring central over, non-central)
    and the boroughs' population sizes.}
}
}
\source{
The present content was prepared using the Detailed
  Gazetteer of the Hungarian Central Statistical Office
  (\url{http://www.ksh.hu/apps/hntr.main?p_lang=EN}). What is
  included in this package is the intellectual product solely of
  this package's authors. Downloaded at 5 April 2019.

The source of the postal codes is Hungarian Postal Service's
(\sQuote{Magyar Posta Zrt.}) website:

\itemize{
  \item{
    \url{https://www.posta.hu/szolgaltatasok/iranyitoszam-kereso},
    downloaded at 5 April 2019)}
  \item{
    \url{https://www.posta.hu/static/internet/download/Allando_postai_szolgaltatohelyek.xlsx},
    downloaded at 12 August 2020}
}
}
\usage{
irsz_2018
}
\description{
A crosswalk table connecting Hungarian Postal Service's postal
codes (\sQuote{\emph{ir}ányító\emph{sz}ám}) and the HCSO's
settlement IDs.
}
\details{
Postal codes include regular codes (used in towns and
  villages), street address based codes (used in cities), and codes
  for post offices (including mobile offices).

Note that this table links all possible postal codes to all settlements using
a particular postal code. To get an unambiguous crosswalk from postal codes
to settlement IDs, use \code{\link{get_irsz_tsz_crosswalk}} to aggregate this
table. As postal codes can cross settlement (and district, county, and even
regional) boundaries, this can lead to some misclassification.
}
\seealso{
\code{\link{get_irsz_tsz_crosswalk}}
}
\keyword{datasets}
