% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teroszt.R
\docType{data}
\name{nav_igazgatosagi_kodok}
\alias{nav_igazgatosagi_kodok}
\title{National Tax and Customs Administration's directorates}
\format{
A tibble with 199 rows, which connects every district
  (\sQuote{járás}) to their respective tax directorate. A fictional
  \code{999} code is used for directorates of Large Taxpayers,
  which opearate on a non-territorial basis.

\describe{
  \item{nav_ig_nev}{Name of the directorate}
  \item{nav_ig_tarsas}{Code of the directorate for corporate
    taxpayers}
  \item{nav_ig_egyeni}{Code of the directorate for sole propriator
    taxpayers}
  \item{megye}{County code, see \link{tsz_2018}}
  \item{jaras}{District code, see \link{tsz_2018}}
}
}
\source{
\url{http://www.nav.gov.hu/nav/adatbazisok/adatbleker/afaalanyok/afaalany_taj.html}.
Downloaded at 10 April 2019.
}
\usage{
nav_igazgatosagi_kodok
}
\description{
A territorial classification used by the National Tax and Customs
Administration (\sQuote{Nemzeti Adó- és Vámhivatal}) to identify
directorates, and a crosswalk between the directorates' counties
(\sQuote{megye}) and districts (\sQuote{járás}) from the
territorial code system.
}
\details{
The directorates' codes correspond the last two digits of
  the tax IDs.
}
\keyword{datasets}
