% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teroszt.R
\docType{package}
\name{teroszt-package}
\alias{teroszt}
\alias{teroszt-package}
\title{teroszt: Data for Hungarian Administrative and Statistical Divisions}
\description{
Data and crosswalk tables for the Central Statistical
     Office's territorial code system
     <http://www.ksh.hu/teruleti_szamjel_menu> and Detailed Gazetteer
     <http://www.ksh.hu/apps/hntr.main?plang=EN>, postal codes
     <https://www.posta.hu/szolgaltatasok/iranyitoszam-kereso>, the
     National Tax and Customs Administration's classifications
     <http://www.nav.gov.hu/nav/adatbazisok/adatbleker/afaalanyok/afaalany_taj.html>
     and OpenStreetMap administrative boundaries
     <https://data2.openstreetmap.hu/hatarok>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/svraka/teroszt}
  \item Report bugs at \url{https://github.com/svraka/teroszt/issues}
}

}
\author{
\strong{Maintainer}: András Svraka \email{svraka.andras@gmail.com}

}
\keyword{internal}
